requirements.txt
azure-storage-blob==12.8.1
dash==1.20.0
plotly==4.14.3
pandas==1.2.4